pyproject.toml
[project]
name = "openai-anthropic-gemini-api-key-check"
version = "0.1.0"
description = "Check API keys for OpenAI, Anthropic, and Gemini by listing models"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "anthropic>=0.75.0",
    "google-genai>=1.56.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[project.scripts]
api-key-check = "main:main"

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
